backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
librosa>=0.10.0
pretty-midi>=0.2.10
pyfluidsynth>=1.3.0
soundfile>=0.12.0
numpy>=1.24.0
scipy>=1.10.0
